Add ZkExponentProof tests
